requirements.txt
# Core Dependencies
crewai>=0.28.0
crewai-tools>=0.4.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0

# GitHub Integration
PyGithub>=1.59.0
gitpython>=3.1.40
requests>=2.31.0
httpx>=0.24.0

# MCP Server
mcp>=0.5.0

# Data Validation and Processing
pydantic>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0
python-dateutil>=2.8.0

# Web Framework
flask>=2.3.0
jinja2>=3.1.0
markupsafe>=2.1.0

# Development Dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# flake8>=6.0.0
# black>=23.0.0
# mypy>=1.0.0
